+++ core/src/db/DBImpl.cpp
@@ -1789,10 +1789,6 @@ DBImpl::CreateIndex(const std::shared_ptr<server::Context>& context, const std::
         // step 3: update index info
         CollectionIndex new_index = index;
         json new_index_json = new_index.extra_params_;
-        if (new_index_json.contains("index_type")) {
-            new_index.engine_type_ = (int32_t)engine::s_map_engine_type.at(new_index_json["index_type"]);
-        }
-
         new_index.metric_type_ = old_index.metric_type_;  // dont change metric type, it was defined by CreateCollection
         if (!utils::IsSameIndex(old_index, new_index)) {
             status = UpdateCollectionIndexRecursively(collection_id, new_index);